Chore: 게시글 신고 ReportType global로 뺌
